apply: move patch data to patch_common.h
