+++ tests/patch/patch_common.h
